third_party/blink/web_tests/wpt_internal/shared_storage/verify-get-undefined.https.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/utils.js"></script>
<script src="/common/utils.js"></script>
<script src="../fenced_frame/resources/utils.js"></script>

<body>
<script>
'use strict';

promise_test(async () => {
  const ancestor_key = token();
  let url0 = generateURL("resources/frame0.html", [ancestor_key]);
  let url1 = generateURL("resources/frame1.html", [ancestor_key]);

  await sharedStorage.worklet.addModule("resources/verify-get-undefined-module.js");

  let uuid1 = await sharedStorage.selectURL(
      "verify-get-undefined-url-selection-operation",
      [{url: url0,
        reportingMetadata: {'click': "resources/frame0.html"}},
       {url: url1,
        reportingMetadata:
          {'mouse interaction': "resources/frame1.html",
          'click': "resources/frame2.html"}}]);

  attachFencedFrame(uuid1, 'opaque-ads');

  // frame1_loaded implies that the operation was successful.
  assert_equals(await nextValueFromServer(ancestor_key), "frame1_loaded");

}, 'verify-get-undefined-url-selection-operation');

</script>
</body>
